Centralise request tracker tests
